Moved PascalCaseToSnakeCase, CamelCaseToSnakeCase, SnakeCaseToPascalCase to StringUtilities
