+++ tests_netcore/Plivo.NetCore.Test/TestClient.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Plivo.Client;
 using Plivo.Http;
-
+using Plivo.Utilities;
 
 namespace Plivo.NetCore.Test
 {
@@ -60,7 +60,7 @@ namespace Plivo.NetCore.Test
     {
         protected override string ResolvePropertyName(string propertyName)
         {
-            return Util.PascalCaseToSnakeCase(propertyName);
+            return StringUtilities.PascalCaseToSnakeCase(propertyName);
         }
     }
 }
